Funcionalidad local y online ya funciona

+++ WhatThePassV2/app/src/main/java/com/uc3m/whatthepass/models/UserRepository.kt
@@ -10,7 +10,7 @@ class UserRepository(private val userDao: UserDao) {
         return userDao.findUserByEmailAndMasterPass(email, password)
     }
 
-    suspend fun readUserEmail(email: String): User {
+    suspend fun readUserByEmail(email: String): User? {
         return userDao.findUserByEmail(email)
     }
 
@@ -26,6 +26,8 @@ class UserRepository(private val userDao: UserDao) {
 
     suspend fun deleteUser(email: String) {
         val user = userDao.findUserByEmail(email)
-        userDao.deleteUser(user)
+        if(user != null) {
+            userDao.deleteUser(user)
+        }
     }
 }
